MVP/requirements.txt
pandas
streamlit
ollama
python-dotenv
plotly
mesop
gradio
openai